Move to where other test commands live
